add messaging to stdout for changes to scheduling (#432)
Return a more helpful message for if the cron job already exists. Previously
returned nothing which may be confusing for a user.

+++ insights/client/schedule.py
@@ -37,6 +37,8 @@ class InsightsSchedule(object):
     def remove_scheduling(self):
         logger.debug('Removing all cron tasks')
         try:
-            os.remove(self.target)
+            if os.path.exists(self.target):
+                os.remove(self.target)
+                return True
         except OSError:
             logger.exception('Could not remove cron.daily')
